Added everything

+++ Blogserver.py
@@ -58,12 +58,7 @@ def make_user_hash(username):
 def verify_user(user_hash):
 	test_user = make_user_hash(user_hash.split("|")[0])
 	if user_hash == test_user:
-
 		return True
-
-
-
-
 
 def verify_login_input(string_input, type):
 	if type == "username":
@@ -78,9 +73,35 @@ def verify_login_input(string_input, type):
 		return "none"
 
 
+#######Database#######
+
+class Comments(db.Model):
+	comment = db.TextProperty(required =True)
+	made_by = db.StringProperty(required=True)
+	post = db.IntegerProperty(required=True)
+	made_date = db.DateTimeProperty(auto_now_add=True)
+
+
+
+class Art(db.Model):
+	title = db.StringProperty(required = True)
+	art = db.TextProperty(required = True)
+	created = db.DateTimeProperty(auto_now_add = True)
+	created_by = db.StringProperty(required = False)
+	vote_points = db.IntegerProperty(required = False, default=0)
+	comment_key = db.ListProperty(db.Key)
+	visible = db.BooleanProperty(default=True)
+	
+
+
+class User(db.Model):
+	username = db.StringProperty(required = True)
+	password = db.StringProperty(required = True)
+	email = db.EmailProperty(required = True)
+	liked_posts = db.ListProperty(int)
+
 
 ########Basic Handler########
-
 
 class Handler(webapp2.RequestHandler):
 
@@ -117,88 +138,84 @@ class Handler(webapp2.RequestHandler):
 		user_record_get = user_record.get()
 		return user_record_get
 
-
-
-#######Database#######
-
-class Comments(db.Model):
-	comment = db.TextProperty(required =True)
-	made_by = db.StringProperty(required=True)
-	post = db.IntegerProperty(required=True)
-	made_date = db.DateTimeProperty(auto_now_add=True)
-
-
-
-class Art(db.Model):
-	title = db.StringProperty(required = True)
-	art = db.TextProperty(required = True)
-	created = db.DateTimeProperty(auto_now_add = True)
-	created_by = db.StringProperty(required = False)
-	vote_points = db.IntegerProperty(required = False, default=0)
-	comment_key = db.ListProperty(db.Key)
-	visible = db.BooleanProperty(default=True)
+	def check_art(self, post_id):
+		check = Art.get_by_id(int(post_id))
+		logging.info(check.key().id())
+		logging.info(post_id)
+	 	return check.key().id() == int(post_id)
+		
+	def check_comments(self, comment_id):
+		check = Comments.get_by_id(int(comment_id))
+		logging.info(self.user)
+		logging.info(check)
+		return check.key().id() == int(comment_id) and check.made_by == self.username()
 	
-
-
-class User(db.Model):
-	username = db.StringProperty(required = True)
-	password = db.StringProperty(required = True)
-	email = db.EmailProperty(required = True)
-	liked_posts = db.ListProperty(int)
-
-
 ########Front Page/Posts########
 
-
 class Mainpage(Handler):
-
+		
 	def save_user_like(self):
-		user_cookie = self.request.cookies.get("user")
 
-		post_id = self.request.get("userlike")
-		post_id = int(post_id)
-		liked_post = Art.get_by_id(int(post_id))
+		if verify_user(self.user):
+			user_cookie = self.request.cookies.get("user")
+			post_id = self.request.get("userlike")
+			if self.check_art(post_id):
+				liked_post = Art.get_by_id(int(post_id))
 
-		user_record = db.GqlQuery("Select * FROM User WHERE username = '%s'" % user_cookie)
-		user_record_get = user_record.get()
+				user_record = db.GqlQuery("Select * FROM User WHERE username = '%s'" % user_cookie)
+				user_record_get = user_record.get()
 
-		for stored_cookie in user_record_get.liked_posts:
-		
-			if post_id == stored_cookie or liked_post.created_by == self.username():
-				break
+				for stored_cookie in user_record_get.liked_posts:
+				
+					if post_id == stored_cookie or liked_post.created_by == self.username():
+						break
+				else:
+
+					user_record_get.liked_posts.append(int(post_id))
+					user_record_get.put()
+					liked_post.vote_points += 1
+					liked_post.put()
+			else:
+				self.redirect("/login")
 		else:
+			self.redirect("/login")
 
-			user_record_get.liked_posts.append(int(post_id))
-			user_record_get.put()
-			
-		
-			liked_post.vote_points += 1
-			liked_post.put()
-		
 	def remove_user_like(self):
-		comment_like_to_unlike_form = self.request.get("userunlike")
-		unliked_post = Art.get_by_id(int(comment_like_to_unlike_form))
-		unliked_post.vote_points = unliked_post.vote_points - 1
-		unliked_post.put()
 
-		logging.error = (self.userrecord())
-		cur_record = self.userrecord()
-		cur_record.liked_posts.remove(int(comment_like_to_unlike_form))
-		cur_record.put()
+		if verify_user(self.user):
+			comment_like_to_unlike_form = self.request.get("userunlike")
+			if self.check_art(comment_like_to_unlike_form):#checks art
+				unliked_post = Art.get_by_id(int(comment_like_to_unlike_form))
+				unliked_post.vote_points = unliked_post.vote_points - 1
+				unliked_post.put()
+
+				logging.error = (self.userrecord())
+				cur_record = self.userrecord()
+				cur_record.liked_posts.remove(int(comment_like_to_unlike_form))
+				cur_record.put()
+			else:
+				self.redirect("/login")
+		else:
+			self.redirect("/login")
 
 
-	def render_login(self, title="", art="", error=""):
+	def render_login(self, title="", art="", error="", *comment):
+
 		arts = db.GqlQuery("Select * FROM Art ORDER BY vote_points DESC")
 		comments = db.GqlQuery("Select * FROM Comments")
 		user_cookie = self.request.cookies.get("user")
 		cur_user = db.GqlQuery("Select * FROM User WHERE username = '%s'" % self.user)
 		userdata = cur_user.get()
-	
-	
-		if verify_user(self.user):
+		if comment:
+			logging.info(comment)
+			comment_to_edit = int(comment[0])
+		else:
+			logging.info(comment) 
+			comment_to_edit = ""
 
-			
-			self.render("mainpage.html",title=title, art=art, error=error, arts=arts, comments=comments, userdata = userdata, username = self.username())
+		if verify_user(self.user):
+			logging.info("fart2 %s", comment_to_edit)
+			self.render("mainpage.html",title=title, art=art, error=error, arts=arts, comments=comments, userdata=userdata, username=self.username(), comment_edit=comment_to_edit)
 		else:
 			self.redirect("/login")
 		
@@ -210,29 +227,70 @@ class Mainpage(Handler):
 		delete_comment = self.request.get("commentdelete")
 		save_like = self.request.get("userlike")
 		save_unlike = self.request.get("userunlike")
+		enable_comment_edit = self.request.get("enableedit")
+		submitedit = self.request.get("submitedit")
+		canceledit = self.request.get("canceledit")
 
-		if delete_comment:
-			comment_to_delete = Comments.get_by_id(int(delete_comment))
-			comment_to_delete.delete()
-			self.redirect("/mainpage")
+		if verify_user(self.user):
 
-		if save_like:
+			if delete_comment:
+				if self.check_art(delete_comment):
+					comment_to_delete = Comments.get_by_id(int(delete_comment))
+					comment_to_delete.delete()
+					self.redirect("/mainpage")
+				else:
+					self.redirect("/mainpage")
 
-			self.save_user_like()
-			self.redirect("/mainpage")
 
-		if save_unlike:
-			self.remove_user_like()
-			self.redirect("/mainpage")
+			if save_like:
 
+				self.save_user_like()
+				self.redirect("/mainpage")
+
+			if save_unlike:
+				self.remove_user_like()
+				self.redirect("/mainpage")
+
+			if enable_comment_edit:
+				logging.info( "fart1 %s", enable_comment_edit)
+				self.render_login('a','b','c', enable_comment_edit) 
+			
+			if submitedit:
+				updated_comment = self.request.get("comment")
+				if self.check_comments(submitedit):
+					comments = Comments.get_by_id(int(submitedit))
+					logging.info("id %s", submitedit)
+					logging.info("comment %s", updated_comment)
+					
+					if not comments:
+						return self.redirect('login')
+					comments.comment = updated_comment
+					comments.put()
+					self.redirect("/mainpage")
+				else:
+					self.error(404)
+
+			if canceledit:
+				self.redirect("/mainpage")
+
+		else:
+			self.redirect("/login")
+
+
+############## PROBLEM HERE END ############################
 
 ########User Login########
 class User_Signup(Handler):
 
  	def get(self):
-		self.render("signup.html")
-		namer = self.request.get("fullname")	
+		if verify_user(self.user):
+			self.response.headers.add_header('Set-Cookie', 'user='';Path=/')
+			self.render("signup.html")
 
+		else:
+
+			self.render("signup.html")
+		
 
 	def post(self):
 		
@@ -266,7 +324,11 @@ class User_Signup(Handler):
 
 class User_Login(Handler):
 	def get(self):
-		self.render("userlogin.html", loginpage = True)
+		if verify_user(self.user):
+			self.response.headers.add_header('Set-Cookie', 'user="";Path=/')
+			self.render("userlogin.html", loginpage = True)
+		else:
+			self.render("userlogin.html", loginpage = True)
 	def post(self):
 		username = self.request.get("username")
 		password = self.request.get("password")
@@ -278,13 +340,9 @@ class User_Login(Handler):
 		if req:
 
 			req_user = req.username
-		
 			req_pass = req.password
 		
 			input_pass_hash = make_pw_hash(username, password, req_pass)
-
-			logging.error ("req_pass = %s and input_pass_hash = %s" % (req_pass, input_pass_hash))
-			#logging.error ("req_user = %s and input_user_hash = %s" % (req_user, input_user_hash))
 
 			if input_pass_hash == req_pass:
 				self.response.headers.add_header('Set-Cookie', 'user=%s;Path=/' % str(req_user))
@@ -298,15 +356,15 @@ class User_Login(Handler):
 			error = "Username Not Found"
 			self.render("userlogin.html", user_error = error)
 
-
-
 class User_Logout(Handler):
 	def get(self):
 		self.render("logout.html")
 		self.response.headers.add_header('Set-Cookie', 'user='';Path=/')
 		time.sleep(3)
 		self.redirect("/login")
-	
+
+
+######Unused Welcome Page#######	
 class Welcome(Handler):
 
 	def get(self):
@@ -324,7 +382,6 @@ class Welcome(Handler):
 		
 class Post(Handler):
 	
-
 	def get(self):
 		if verify_user(self.user):
 			self.render("post.html", username = self.username())
@@ -335,44 +392,57 @@ class Post(Handler):
 		current_user = self.request.cookies.get("user").split("|")[0]
 		title = self.request.get("title")
 		art = self.request.get("art")
-
-		
-
 		if title and art:
-			a = Art(title = title, art = art, created_by = current_user)
-			a.put()
-			time.sleep(1)
-			self.redirect("/postview/%s" % a.key().id())
-		
+
+			if verify_user: #user verified
+				a = Art(title = title, art = art, created_by = current_user)
+				a.put()
+				time.sleep(1)
+				self.redirect("/postview/%s" % a.key().id())
+			else:
+				self.redirect("/login")	
+			
 		else:
-			error = "Fill it all out dumbwag"
+			error = "Needs a title and a post to continue"
 			self.render("post.html", title=title, art=art, error=error, username = self.username())
 
 class PostHandler(Handler):
+
 	def get(self, post_id):
-		if verify_user(self.user):
+
+		if Art.get_by_id(int(post_id)): #verified art
 			arts = Art.get_by_id(int(post_id))
-			self.render("postview.html", arts=[arts], username = self.username())
+
+			if verify_user(self.user): #verified user
+				self.render("postview.html", arts=[arts], username = self.username())
+			else:
+				self.redirect("/login")
 		else:
-			self.redirect("/login")
+			self.redirect("/mainpage")
 	
 	def post(self, post_id):
+
+
 		comment = self.request.get("comment")
 
-		post = Art.get_by_id(int(post_id))
-		new_comment = Comments(comment = comment, made_by = self.username(), post = int(post_id))
-		new_comment.put()
-		post.comment_key.append(new_comment.key())
-		post.put()
-	
-		self.redirect("/mainpage")
+		if verify_user(self.user) and self.check_art(post_id): #verified user and art
+			post = Art.get_by_id(int(post_id))
+			new_comment = Comments(comment = comment, made_by = self.username(), post = int(post_id))
+			new_comment.put()
+			post.comment_key.append(new_comment.key())
+			post.put()
+			self.redirect("/mainpage")
+		else:
+			logging.error(post_id)
+			logging.error(self.check_art(post_id))
+			self.error(404)
 
 
 class PostEdit(Handler):
 	def get(self, post_id):
-		if verify_user(self.user):
+		if verify_user(self.user): #verified user
 			arts = Art.get_by_id(int(post_id))
-			if self.username() == arts.created_by:
+			if self.username() == arts.created_by: #verified owner
 				
 				self.render("postedit.html", arts=[arts], username = self.username())
 			else:
@@ -384,29 +454,33 @@ class PostEdit(Handler):
 
 	def post(self, post_id):
 	
+		if verify_user(self.user):#verified user
+			delete = self.request.get("deletebutt", None)
+			edit = self.request.get("submitedit", None)
+			
+			if delete and check_art(post_id): #verified art exists
+				post_to_edit = Art.get_by_id(int(post_id))
+				if post_to_edit.made_by == self.username():
+					post_to_edit.visible = False
+					post_to_edit.put()
+					self.redirect("/mainpage")
+				else:
+					self.redirect("/login")
 
-		
-		
-		delete = self.request.get("deletebutt", None)
-		edit = self.request.get("submitedit", None)
-		
-		if delete:
-			post_to_edit = Art.get_by_id(int(post_id))
-			post_to_edit.visible = False
-			post_to_edit.put()
-			self.redirect("/mainpage")
-	
-		if edit:
-			post_to_edit = Art.get_by_id(int(post_id))
-			content = self.request.get("post-body")
-			post_to_edit.art = content
-			post_to_edit.put()
-			time.sleep(0.5)
-			self.redirect("/mainpage")
+			if edit and check_art(post_id): #verified art exists
+				post_to_edit = Art.get_by_id(int(post_id))
+				if post_to_edit.made_by == self.username():
+					content = self.request.get("post-body")
+					post_to_edit.art = content
+					post_to_edit.put()
+					time.sleep(0.5)
+					self.redirect("/mainpage")
+				else:
+					self.redirect("/login")
+		else:
+			self.redirect("/login")
 
-
-
-########
+########Sets cookie for whole site#######
 
 class Toplevel(Handler):
 	def get(self):
@@ -425,8 +499,6 @@ class Toplevel(Handler):
  			self.redirect("/mainpage")
  		else:
  			self.redirect("/mainpage")
-
-
 
 
 app = webapp2.WSGIApplication ([("/", Toplevel),("/mainpage", Mainpage), ("/post", Post), (r"/postview/(\d+)", PostHandler),(r"/postedit/(\d+)", PostEdit), ("/signup", User_Signup), ("/welcome", Welcome), ("/login", User_Login), ("/logout", User_Logout)] , debug=True)
